Move exr_utils.py to utils module and add test
Previously it was only usable from the safetensors_generator script, but this makes it usable more widely
in the repo. This will be used in a following patch for evaluating alternative result sets.

Functions which are not generic EXR functions (validate_exr_dataset_structure and create_depth_params,
which are specific to our dataset format) have been moved into dataset_reader.py directly.
