+++ src/ng_model_gym/utils/exr_utils.py
@@ -2,34 +2,11 @@
 # its affiliates <open-source-office@arm.com></text>
 # SPDX-License-Identifier: Apache-2.0
 from pathlib import Path
-from typing import Dict, Union
+from typing import Union
 
 import numpy as np
 import OpenEXR
 import torch
-
-from scripts.safetensors_generator.fsr2_methods import depth_to_view_space_params
-
-
-def validate_exr_dataset_structure(
-    src_root: Path, seq_path: Path, scale: float
-) -> None:
-    """Check the dataset is in the expected format."""
-    scale_str = str(float(scale)).replace(".", "_").replace("_0", "")
-
-    required_dirs = [
-        src_root / "ground_truth" / seq_path,
-        src_root / "motion_gt" / seq_path,
-        src_root / f"x{scale_str}" / "color" / seq_path,
-        src_root / f"x{scale_str}" / "depth" / seq_path,
-        src_root / f"x{scale_str}" / "motion" / seq_path,
-    ]
-
-    for required_dir in required_dirs:
-        if not required_dir.is_dir():
-            raise FileNotFoundError(f"Missing {required_dir} in expected structure.")
-        if not any(required_dir.iterdir()):
-            raise ValueError(f"{required_dir} is empty.")
 
 
 def read_exr(
@@ -108,18 +85,3 @@ def pixel_type_to_dtype_and_bytes(ptype: OpenEXR.PixelType) -> tuple[np.dtype, i
     if pval == OpenEXR.UINT:
         return np.uint32, 4
     raise ValueError(f"Unsupported pixel type: {ptype}")
-
-
-def create_depth_params(data: Dict[str, torch.Tensor]) -> Dict[str, torch.Tensor]:
-    """Create the depth parameters"""
-    make_image_like = lambda t: t.unsqueeze(-1).unsqueeze(-1)
-    depth_params = depth_to_view_space_params(
-        zNear=make_image_like(data["zNear"]),
-        zFar=make_image_like(data["zFar"]),
-        FovY=make_image_like(data["FovY"]),
-        infinite=make_image_like(data["infinite_zFar"]).to(torch.bool),
-        renderSizeWidth=make_image_like(data["render_size"])[:, 1:2, ...],
-        renderSizeHeight=make_image_like(data["render_size"])[:, 0:1, ...],
-        inverted=torch.tensor(False, dtype=torch.bool),
-    ).squeeze()
-    return depth_params
